requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
pandas>=1.3.0
requests>=2.25.0
pytesseract>=0.3.8
Pillow>=8.0.0
scikit-learn>=0.24.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
matplotlib>=3.4.0
flask>=2.0.0